Using long for timestamps

+++ src/main/java/me/leoko/advancedban/manager/DatabaseManager.java
@@ -300,8 +300,8 @@ public class DatabaseManager {
                             result.getString("reason"),
                             result.getString("operator"),
                             result.getString("punishmentType"),
-                            result.getInt("start"),
-                            result.getInt("end"),
+                            result.getLong("start"),
+                            result.getLong("end"),
                             result.getString("calculation")
                     );
